backend/requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pg8000==1.30.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# File handling
aiofiles==23.2.1

# Utils
python-dateutil==2.8.2
phonenumbers==8.13.27

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
